Predict the COI

Predict the COI of the sample by comparing the within sample allele frequency (WSAF) and the population level allele frequency (PLAF) of the sample to what a theoretical WSAF and PLAF should look like. By examining the sample's WSAF and PLAF to the theoretical WSAF and PLAF, an estimation can be made about what the COI of the sample is. We refer to the sample's WSAF vs PLAF as the "sample curve" and refer to the theoretical WSAF vs PLAF as the "theoretical curve." To determine the predicted COI value, one of three different methods can be selected:

end

Determines the distance between the theoretical and sample curve at a PLAF of 0.5. The COI is whichever theoretical COI curve has the smallest distance to the simulated data.

ideal

Determines the distance between the theoretical and sample curve at the ideal PLAF. The ideal PLAF is calculated by looking at the change between the COI of \(i\) and the COI of \(i-1\) and finding the PLAF for which this distance is maximized. The COI is whichever theoretical COI curve has the smallest distance to the simulated data at the ideal PLAF.

overall

Determines the distance between the theoretical and simulated curve for all PLAFs. Computes the distance between the theoretical curves and the simulated curve. The COI is whichever theoretical curve has the smallest distance to the simulated curve. There is an option to choose one of several distance metrics:

abs_sum: Absolute value of sum of difference.

sum_abs: Sum of absolute difference.

squared: Sum of squared difference.

KL: Kullback-Leibler divergence.

compute_coi(
  processed_data,
  theory_coi_range,
  cut,
  method = "overall",
  dist_method = "squared",
  weighted = TRUE
)

Arguments

processed_data
The processed COI data. This is the output of process_simulated_coi or process_real_data.

theory_coi_range
The range of COIs for which theoretical curves will be generated.

cut
A vector indicating how often the data is summarized.

method
The method to be employed. One of "end", "ideal", "overall".

dist_method
The distance method used to determine the distance between the theoretical and simulated curves for the "overall" method. One of "abs_sum", "sum_abs", "squared", "KL".

weighted
An indicator indicating whether to compute the weighted distance.

Value

A list of the following:

coi

The predicted COI of the sample.

probability

A probability density function representing the probability of each COI.