Likelihood of a COI

A function to generate the likelihood of a specific COI value. The likelihood can be thought of the distance between two curves: the "real" COI curve, generated from the inputted data, and the "simulated" COI curve, which depends on the COI value specified.

likelihood(coi, processed_data, dist_method = "squared", weighted = TRUE)

Arguments

coi
The COI for which the likelihood will be generated.

processed_data
The processed COI data. This is the output of process_sim() or process_real().

dist_method
The distance method used to determine the distance between the theoretical and simulated curves for the "overall" method. One of "abs_sum", "sum_abs", "squared".

weighted
An indicator indicating whether to compute the weighted distance.

Value

The likelihood for a specific COI value.