Process real data

Generate the COI curve for real data.

process_real(wsaf, plaf, seq_error = NULL, bin_size = 20, coi_method = "1")

Arguments

wsaf
The within-sample allele frequency.

plaf
The population-level allele frequency.

seq_error
The level of sequencing error that is assumed. If no value is inputted, then we infer the level of sequence error.

bin_size
The minimum size of each bin of data.

coi_method
The method we will use to calculate the theoretical COI. The method is either "1" or "2". The default value is "1".

Value

A list of the following:

data: A tibble with

plaf_cut: Breaks of the form [a, b).

m_variant: The average WSAF or proportion of variant sites in each segment defined by plaf_cut.

bucket_size: The number of loci in each bucket.

midpoints: The midpoint of each bucket.

seq_error: The sequence error inferred.

bin_size: The minimum size of each bin.

cuts: The breaks utilized in splitting the data. of each COI.

Details

The function computes whether a SNP is a variant site or not, based on the simulated WSAF at that SNP. This process additionally accounts for potential sequencing error.

See also