Simulate biallelic data

Simulate biallelic data from a simple statistical model. Inputs include the complexity of infection (COI), population-level allele frequencies (PLAF), and some parameters dictating skew and error distributions. Outputs include the phased haplotypes and the unphased read count and coverage data.

sim_biallelic(
  coi = 3,
  plaf = runif(10, 0, 0.5),
  coverage = 200,
  alpha = 1,
  overdispersion = 0,
  relatedness = 0,
  epsilon = 0
)

Arguments

coi
Complexity of infection.

plaf
Vector of population-level allele frequencies at each locus.

coverage
Coverage at each locus. If a single value then the same coverage is applied over all loci.

alpha
Shape parameter of the symmetric Dirichlet prior on strain proportions.

overdispersion
The extent to which counts are over-dispersed relative to the binomial distribution. Counts are Beta-binomially distributed, with the beta distribution having shape parameters \(\frac{p}{overdispersion}\) and \(\frac{1-p}{overdispersion}\).

relatedness
The probability that a strain in mixed infections is related to another. Default = 0 (unrelated). The implementation is similar to relatedness as defined in THE REAL McCOIL simulations. In the original paper (https://doi.org/10.1371/journal.pcbi.1005348) this is defined as: "... simulated relatedness (r) among lineages within the same host by sampling alleles either from an existing lineage within the same host (with probability r) or from the population (with probability (1-r))."

epsilon
The probability of a single read being miscalled as the other allele. Applies in both directions.

Value

A list of:

coi: The COI used to simulate the data.

strain_proportions: The strain proportion of each strain.

phased: The phased haplotype.

data: A dataframe of:

plaf: The population-level allele frequency.

coverage: The coverage at each locus.

counts: The count at each locus.

wsaf: The within-sample allele frequency.

inputs: A dataframe of function input arguments:

alpha: Shape parameters of Dirichlet controlling strain proportions.

overdispersion: Overdispersion in count data.

relatedness: Within sample relatedness between strains.

epsilon: Probability of a single read being miscalled.

Details

Simulated data are drawn from a simple statistical model:

Strain proportions are drawn from a symmetric Dirichlet distribution with shape parameter alpha.

Phased haplotypes are drawn at every locus, one for each coi. The allele at each locus is drawn from a Bernoulli distribution with probability given by the plaf.

The "true" within-sample allele frequency at every locus is obtained by multiplying haplotypes by their strain proportions, and summing over haplotypes. Errors are introduced through the equation \[wsaf_{error} = wsaf(1-e) + (1-wsaf)e\] where \(wsaf\) is the WSAF without error and \(e\) is the error parameter epsilon.

Final read counts are drawn from a beta-binomial distribution with expectation \(w_{error}\). The raw number of draws is given by the coverage, and the skew of the distribution is given by the overdispersion parameter. If the overdispersion is equal to zero, then the distribution is binomial, rather than beta-binomial.

See also