Single sensitivity analysis

Runs a single full COI sensitivity analysis.

single_sensitivity(
  COI = 3,
  max_COI = 25,
  PLAF = runif(1000, 0, 0.5),
  coverage = 200,
  alpha = 1,
  overdispersion = 0,
  epsilon = 0,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  method = "overall",
  dist_method = "squared",
  weighted = TRUE
)

Arguments

COI
Complexity of infection.

max_COI
A number indicating the maximum COI to compare the simulated data to.

PLAF
Vector of population-level allele frequencies at each locus.

coverage
Coverage at each locus. If a single value then the same coverage is applied over all loci.

alpha
Shape parameter of the symmetric Dirichlet prior on strain proportions.

overdispersion
The extent to which counts are over-dispersed relative to the binomial distribution. Counts are Beta-binomially distributed, with the beta distribution having shape parameters \(\frac{p}{overdispersion}\) and \(\frac{1-p}{overdispersion}\).

epsilon
The probability of a single read being miscalled as the other allele. Applies in both directions.

seq_error
The level of sequencing error that is assumed.

cut
A vector indicating how often the data is summarized.

method
The method to be employed. One of "end", "ideal", "overall".

dist_method
The distance method used to determine the distance between the theoretical and simulated curves for the "overall" method. One of "abs_sum", "sum_abs", "squared".

weighted
An indicator indicating whether to compute the weighted distance.

Value

Predicted COI value.