Predict the COI

Predict the COI of the sample.

compute_coi(
  data,
  data_type,
  max_coi = 25,
  seq_error = NULL,
  bin_size = 20,
  comparison = "overall",
  distance = "squared",
  coi_method = "1"
)

Arguments

data
The data for which the COI will be computed.

data_type
The type of the data to be analyzed. One of "sim" or "real".

max_coi
A number indicating the maximum COI to compare the simulated data to.

seq_error
The level of sequencing error that is assumed. If no value is inputted, then we infer the level of sequence error.

bin_size
The minimum size of each bin of data.

comparison
The method to be employed. One of "end", "ideal", "overall".

distance
The distance method used to determine the distance between the theoretical and simulated curves for the "overall" method. One of "abs_sum", "sum_abs", "squared".

coi_method
The method we will use to calculate the theoretical COI. The method is either "1" or "2". The default value is "1".

Value

A list of the following:

coi: The predicted COI of the sample.

probability: A probability density function representing the probability of each COI.

Details

Compare the within sample allele frequency (WSAF) and the population level allele frequency (PLAF) of the sample to what a theoretical WSAF and PLAF should look like. By examining the sample's WSAF and PLAF to the theoretical WSAF and PLAF, an estimation can be made about what the COI of the sample is. We refer to the sample's WSAF vs PLAF as the "sample curve" and refer to the theoretical WSAF vs PLAF as the "theoretical curve." To determine the predicted COI value, one of three different methods can be selected:

end

Determines the distance between the theoretical and sample curve at a PLAF of 0.5. The COI is whichever theoretical COI curve has the smallest distance to the simulated data.

ideal

Determines the distance between the theoretical and sample curve at the ideal PLAF. The ideal PLAF is calculated by looking at the change between the COI of \(i\) and the COI of \(i-1\) and finding the PLAF for which this distance is maximized. The COI is whichever theoretical COI curve has the smallest distance to the simulated data at the ideal PLAF.

overall

Determines the distance between the theoretical and simulated curve for all PLAFs. Computes the distance between the theoretical curves and the simulated curve. The COI is whichever theoretical curve has the smallest distance to the simulated curve. There is an option to choose one of several distance metrics:

abs_sum: Absolute value of sum of difference.

sum_abs: Sum of absolute difference.

squared: Sum of squared difference.