Compute distance between two curves

Compute the distance between two curves using several methods.

abs_sum

Absolute value of sum of difference.

sum_abs

Sum of absolute difference.

squared

Sum of squared difference.

Usage

distance_curves(processed_data, theory_cois, distance = "squared")

Arguments

theory_cois

The theoretical COI curves.

distance

The distance method used to determine the distance between the theoretical and simulated curves for the "overall" method. One of "abs_sum", "sum_abs", "squared".

Value

A list of the following:

coi

The predicted COI for the simulation.

dist

The distance between each theoretical COI and the simulated COI curve.