Optimize the COI

A function to compute the COI of inputted data.

Usage,

optimize_coi(
  data,
  data_type,
  max_coi = 25,
  seq_error = NULL,
  bin_size = 20,
  distance = "squared",
  coi_method = "1"
)

Arguments

data

The data for which the COI will be computed.

data_type

The type of the data to be analyzed. One of "sim" or "real".

max_coi

A number indicating the maximum COI to compare the simulated data to.

seq_error

The level of sequencing error that is assumed. If no value is inputted, then we infer the level of sequence error.

bin_size

The minimum size of each bin of data.

distance

The distance method used to determine the distance between the theoretical and simulated curves for the "overall" method. One of "abs_sum", "sum_abs", "squared".

coi_method

The method we will use to calculate the theoretical COI. The method is either "1" or "2". The default value is "1".

Value

The predicted COI value.

Details

The function utilizes stats::optim() In particular, the function utilizes a quasi-Newton method to compute gradients and build a picture of the surface to be optimized. The function uses a likelihood function as defined by likelihood().

See also